+++ src/Entities/Types/Identity/Uuid.php
@@ -2,37 +2,13 @@
 
 namespace Somnambulist\Domain\Entities\Types\Identity;
 
-use Assert\Assert;
-use Somnambulist\Domain\Entities\AbstractValueObject;
-
 /**
  * Class Uuid
  *
  * @package    Somnambulist\Domain\Entities\Types\Identity
  * @subpackage Somnambulist\Domain\Entities\Types\Identity\Uuid
  */
-class Uuid extends AbstractValueObject
+final class Uuid extends AbstractIdentity
 {
 
-    /**
-     * @var string
-     */
-    private $value;
-
-    /**
-     * Constructor.
-     *
-     * @param string $uuid
-     */
-    public function __construct(string $uuid)
-    {
-        Assert::that($uuid, null, 'uuid')->notEmpty()->uuid();
-
-        $this->value = $uuid;
-    }
-
-    public function toString(): string
-    {
-        return (string)$this->value;
-    }
 }